// yugabyted-ui/ui/src/api/src/models/TaskTypeEnum.ts
// tslint:disable
/**
 * Yugabyte Cloud
 * YugabyteDB as a Service
 *
 * The version of the OpenAPI document: v1
 * Contact: support@yugabyte.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * The type of task
 * @export
 * @enum {string}
 */
export enum TaskTypeEnum {
  CreateCluster = 'CREATE_CLUSTER',
  EditCluster = 'EDIT_CLUSTER',
  DeleteCluster = 'DELETE_CLUSTER',
  EditAllowList = 'EDIT_ALLOW_LIST',
  CreateBackup = 'CREATE_BACKUP',
  RestoreBackup = 'RESTORE_BACKUP',
  MigrateCluster = 'MIGRATE_CLUSTER',
  DeleteProject = 'DELETE_PROJECT',
  UpgradeCluster = 'UPGRADE_CLUSTER',
  PauseCluster = 'PAUSE_CLUSTER',
  ResumeCluster = 'RESUME_CLUSTER',
  SetupXclusterReplication = 'SETUP_XCLUSTER_REPLICATION'
}